[Profiler] Memory profiler part 6: Mark gradients and temporary intermediates.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87566

Semantic assignment will be built up as a series of passes which gradually pin down the regions of a trace. For this reason it is important to be very meticulous in the assignment of categories.

We begin with gradients as they are both straightforward to identify and foundational to subsequent analysis. There are two mechanisms that the profiler can use to tag gradients, each with their own advantages and limitations. The first is direct inspection of the op graph which is generic but predicated on certain features of the Autograd engine. (And therefore not necessarily exhaustive.) The second approach is direct instrumentation via the python tracer. This method relies requires that gradients be attached to an nn.Module parameter and can miss corner cases such as `set_to_none=True` due to the cache structure of the python tracer. Combined these two approaches provide very high coverage.

Temporaries are more straightforward; we can easily add them by trivial local inspection of a data flow node.

Because this is the first PR in the end-to-end section most of the code is building the scaffolding for category bookkeeping and unit testing. (The actual gradient extraction was covered in an earlier PR.)
ghstack-source-id: 174078149

Differential Revision: [D40220389](https://our.internmc.facebook.com/intern/diff/D40220389/)