[Profiler] Memory profiler part 7: Mark inputs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87567

It is surprisingly difficult to identify the leaves of the data flow graph. The issue is that inputs and pre-existing parameters look identical until parameter identification takes place. It's not too bad for training since Autograd lets us differentiate between them however I still want the tool to do something reasonable in inference.

Some of this will be ameliorated when a later PR pulls in parameters from python tracing. The current approach is passable, but I will continue to mull over refinements.
ghstack-source-id: 174078148

Differential Revision: [D40220388](https://our.internmc.facebook.com/intern/diff/D40220388/)