Allow Module forward-pre and forward hooks to take kwargs
closes #35643

This PR is mostly copied from #82042. Thanks Padarn for implementing
the first version and debugging into the errors.

Based on the discussion in #82042, this PR adds a `with_kwargs`
argument to `register_forward_pre_hook` and `register_forward_hook`
methods. When the arg is set to true, the provided hook must accept
kwargs args. Under the hood, the hook is wrapped into `_ForwardHook`
and `_ForwardPreHook` types to avoid backward compatibility issues.

ghstack-source-id: 3356bd648b02a252ffbd038f16d69d4c213a74cd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89389